Rename priceCalculatorAtoms -> priceIntentAtoms
