+++ apps/store/src/components/PriceCalculator/PetDogBreedsField.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'next-i18next'
 import { useState, useMemo } from 'react'
 import { Space } from 'ui'
 import { Combobox } from '@/components/Combobox/Combobox'
-import { priceCalculatorLoadingAtom } from '@/components/PriceCalculator/priceCalculatorAtoms'
+import { priceCalculatorLoadingAtom } from '@/components/ProductPage/PurchaseForm/priceIntentAtoms'
 import { PriceIntentAnimal, usePriceIntentAvailableBreedsQuery } from '@/services/graphql/generated'
 import type {
   PetDogBreedsField as InputFieldPetDogBreeds,
