+++ apps/store/src/components/PriceCalculator/PriceCalculatorSection.tsx
@@ -4,7 +4,7 @@ import NextLink from 'next/link'
 import { useTranslation } from 'next-i18next'
 import { type FormEventHandler, type ReactNode } from 'react'
 import { Button, Space, Text } from 'ui'
-import { priceCalculatorLoadingAtom } from '@/components/PriceCalculator/priceCalculatorAtoms'
+import { priceCalculatorLoadingAtom } from '@/components/ProductPage/PurchaseForm/priceIntentAtoms'
 import { linkStyles } from '@/components/RichText/RichText.styles'
 import { deserializeField } from '@/services/PriceCalculator/PriceCalculator.helpers'
 import { type FormSection, type JSONData } from '@/services/PriceCalculator/PriceCalculator.types'
