+++ apps/store/src/components/PriceCalculator/SsnSeSection.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'next-i18next'
 import { type FormEventHandler, memo } from 'react'
 import { Button, Space } from 'ui'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
-import { shopSessionCustomerAtom } from '@/components/PriceCalculator/priceCalculatorAtoms'
+import { shopSessionCustomerAtom } from '@/components/ProductPage/PurchaseForm/priceIntentAtoms'
 import { useShopSessionCustomerUpdateMutation } from '@/services/graphql/generated'
 import { useShopSessionId } from '@/services/shopSession/ShopSessionContext'
 import { useErrorMessage } from '@/utils/useErrorMessage'
